# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservice with auth and notes of user and borrowed books",
        "title": "User Book microservice API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Dinmikhamed Amurzakov",
            "email": "dimash.amurzakov@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "paths": {
...
        "/update-pass": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "UpdatePassword",
                "operationId": "update-password",
                "parameters": [
                    {
                        "description": "current and new password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdatePasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "description": "create user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "CreateUser",
                "operationId": "create-account",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/delete": {
            "get": {
                "description": "delete user by username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "DeleteUser",
                "operationId": "delete-user-by-username",
                "responses": {
                    "200": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/get_by_id": {
            "get": {
                "description": "get users by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "GetUserByID",
                "operationId": "get-users-by_id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/get_by_name": {
            "get": {
                "description": "get users by username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "GetUserByUsername",
                "operationId": "get-users-by-username",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID, Name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.UpdatePasswordReq": {
            "type": "object",
            "properties": {
                "current_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object"
        },
        "models.UserAuth": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}